`RegularExpressionS` を拡張機能へ移動
