+++ src/Covid19Radar.LogViewer/SearchFilters/MessageSearchFilters.cs
@@ -105,30 +105,5 @@ namespace Covid19Radar.LogViewer.SearchFilters
 				}
 			}
 		}
-
-		internal sealed class RegularExpressionS : ISearchFilter
-		{
-			internal static readonly RegularExpressionS _inst = new();
-
-			public string Key => "regexs";
-
-			private RegularExpressionS() { }
-
-			public bool Match(SearchFilterNode value, LogDataModel model)
-			{
-				if (value is SearchFilterNodeList nodeList && nodeList.Nodes.Count >= 2 &&
-					nodeList.Nodes[0] is TokenNode patternNode) {
-					string   pattern  = patternNode.Token.GetText();
-					string[] expected = new string[nodeList.Nodes.Count - 1];
-					for (int i = 0; i < expected.Length; ++i) {
-						expected[i] = nodeList.Nodes[i + 1] is TokenNode node ? node.Token.GetText() : string.Empty;
-					}
-					return expected.SequenceEqual(Regex.Split(model.OriginalMessage,    pattern))
-						|| expected.SequenceEqual(Regex.Split(model.TransformedMessage, pattern));
-				} else {
-					return false;
-				}
-			}
-		}
 	}
 }
